Move callback interfaces into way segment parser
