--- core/src/main/java/com/graphhopper/reader/osm/EdgeHandler.java
@@ -1,28 +1,0 @@
-/*
- *  Licensed to GraphHopper GmbH under one or more contributor
- *  license agreements. See the NOTICE file distributed with this work for
- *  additional information regarding copyright ownership.
- *
- *  GraphHopper GmbH licenses this file to you under the Apache License,
- *  Version 2.0 (the "License"); you may not use this file except in
- *  compliance with the License. You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package com.graphhopper.reader.osm;
-
-import com.graphhopper.reader.ReaderWay;
-import com.graphhopper.util.PointList;
-
-import java.util.Map;
-
-interface EdgeHandler {
-    void handleEdge(int from, int to, PointList pointList, ReaderWay way, Map<String, Object> nodeTags);
-}
